refactor: move XSD download button from header to footer (#388)
* refactor: move XSD download button from header to footer

Remove the download XSD button and related code from the header component,
including its styles and downloadXSD() method. Add the download XSD button
to the footer component instead, implementing the downloadXSD() method there.

This change declutters the header by removing an underused action and places
the download functionality in the footer where it is more accessible and
contextually appropriate. It also cleans up unused styles and imports in the
header component.

* fix: center footer icons